+++ Aplicatie gestionare zboruri companie/backend/src/main/java/com/flights/account/service/ShoppingCartService.java
@@ -1,5 +1,6 @@
 package com.flights.account.service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,7 +27,8 @@ public class ShoppingCartService {
 	}
 	
 	public ShoppingCart findByUserEmailAndTicketId(String userEmail, int ticketId) {
-        return shoppingCartRepository.findByUserEmailAndTicketId(userEmail, ticketId);
+		String status = "not ordered";
+        return shoppingCartRepository.findByUserEmailAndTicketIdAndStatus(userEmail, ticketId, status);
     }
 	
 	
@@ -35,14 +37,53 @@ public class ShoppingCartService {
     	
     	for(ShoppingCart product : shoppingCartProducts) {
     		if(product.getUser().getEmail().equals(shoppingCartDto.getUserEmail()) &&
-    				product.getTicket().getId().equals(shoppingCartDto.getTicketId())){
+    				product.getTicket().getId().equals(shoppingCartDto.getTicketId()) &&
+    				product.getStatus().equals("not ordered")){
     			shoppingCartRepository.delete(product);
     		}
     	}
 		
     }
 	
+	public void buyTicketsFromUserShoppingCart(ShoppingCartDto shoppingCartDto) {
+    	List<ShoppingCart> shoppingCartProducts = shoppingCartRepository.findAll();
+    	
+    	for(ShoppingCart product : shoppingCartProducts) {
+    		if(product.getUser().getEmail().equals(shoppingCartDto.getUserEmail())){
+    			product.setStatus("ordered");
+    			shoppingCartRepository.save(product);
+    		}
+    	}
+		
+    }
+	
 	public List<ShoppingCart> getFullShoppingCart() {
-        return shoppingCartRepository.findAll();
+		List<ShoppingCart> fullShoppingCart = shoppingCartRepository.findAll();
+        List<ShoppingCart> shoppingCart = new ArrayList<>();
+        
+        if(fullShoppingCart != null) {
+            for(ShoppingCart product : shoppingCartRepository.findAll()) {
+            	if(product.getStatus().equals("not ordered")) {
+            		shoppingCart.add(product);
+            	}
+            }
+        }
+
+        return shoppingCart;
+    }
+	
+	public List<ShoppingCart> getHistoryTransaction() {
+		List<ShoppingCart> fullShoppingCart = shoppingCartRepository.findAll();
+        List<ShoppingCart> shoppingCart = new ArrayList<>();
+        
+        if(fullShoppingCart != null) {
+            for(ShoppingCart product : shoppingCartRepository.findAll()) {
+            	if(product.getStatus().equals("ordered")) {
+            		shoppingCart.add(product);
+            	}
+            }
+        }
+
+        return shoppingCart;
     }
 }
